Updated swim_speed_module.py
When switching users or creating users, index out of range error would get thrown. Improper file saving would occur as well. These were all fixed as well as a tiny change to the calibration. Since calibration is mandatory prior to swim speed test, the need for calibration as an option on the main menu is no longer needed. Decided to omit that option on the main menu.